requirements.txt
pandas==1.3.3
numpy==1.19.5
matplotlib==3.4.3
holidays==0.11.2
python-dateutil==2.8.2
statsmodels==0.12.2
tensorflow==2.6.0
sklearn==0.0
scikit-learn==0.24.2
requests==2.25.1
lxml==4.6.3
jieba==0.42.1
pymsteams==0.1.16
urllib3==1.26.6
pytz==2021.1
bs4==0.0.1
beautifulsoup4==4.10.0
selenium==3.141.0
yfinance==0.1.63